finish usecase finish usecase

+++ domain/src/main/java/com/example/domain/usercases/DeleteDeviceFromDatabaseByID.kt
@@ -4,5 +4,5 @@ import Resource
 import com.example.domain.repo.UserRepo
 
 class DeleteDeviceFromDatabaseByID(private val userRepo: UserRepo) {
-    suspend operator fun invoke(id:Int?,result: (Resource<String>) -> Unit)=userRepo.deleteDeviceFromDatabaseByID(id,result)
+    suspend operator fun invoke(id:Long?,result: (Resource<String>) -> Unit)=userRepo.deleteDeviceFromDatabaseByID(id,result)
 }
